--- diffusion-models/conf/hydra/simple.yaml
@@ -1,14 +1,0 @@
-run:
-  dir: .
-
-output_subdir: null
-
-job_logging:
-  version: 1
-  handlers:
-    console:
-      class: logging.StreamHandler
-      stream: ext://sys.stdout
-  root:
-    handlers: [console]
-  disable_existing_loggers: false
