+++ diffusion-models/config/yaml/config_generate.yaml
