+++ diffusion-models/config/yaml/dataset/imagenet.yaml
