+++ diffusion-models/config/yaml/diffusion/cosine.yaml
