+++ diffusion-models/config/yaml/diffusion/linear.yaml
