+++ diffusion-models/config/yaml/hydra/job_logging/custom.yaml
@@ -1,0 +1,16 @@
+version: 1
+
+formatters:
+  simple:
+    format: '%(asctime)s [%(name)s][%(levelname)s] %(message)s'
+
+handlers:
+  console:
+    class: logging.StreamHandler
+    formatter: simple
+    stream: ext://sys.stdout
+
+root:
+  handlers: [console]
+
+disable_existing_loggers: false
