+++ diffusion-models/config/yaml/model/imagen_64.yaml
