+++ diffusion-models/config/yaml/train/imagenet_64to256_efficient_unet.yaml
